backend/requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
firebase-admin==6.1.0
pytesseract==0.3.10
Pillow==9.5.0
google-generativeai==0.3.2